Add files via upload
Design of university unit selection system

The university unit selection system determines the courses and the number of units that can be offered in the new semester according to the courses passed in the previous semester as well as the average of the courses of the previous semester. This system notifies the user of the necessary messages and warnings in case of selecting courses that overlap with each other or do not observe the prerequisite and required courses. After selecting the unit, this system determines the semester tuition.